Dynamic table load
Got the table up and running with a little help (thanks you know who you
are). Was not impressed with using the native Jquery so switched to JS
and then used Jquery to html instead of append. There is probably a more
jquery way of doing that but hey that's why you use Jquery write what
you want to get to do what you want. FU Sencha