Season 4 champs
Adding in the ability to only do stats off of Season 4 supports. TODO.
Condense the JS functions and GD functions so that we only have to call
one function to get data. Right now it is silly having the same
functionality copied to do the same thing over and over. Right now
trying to get the functionality first then condense the code to make
more sense.