Removing Dead nodes
With Season 6 the ward purchase is not applying to Pink wards and Green
wards were removed from the game. I am removing that node as it is
basically pointless. In the future I might remove it from the
collector. My hope is that riot sees their mistake and adds that back.
The Map name has never been that great and Riot keeps changing the
names of the maps so it is useless. Kept Mode as it is more telling
anyways.