Champ Game Stats
Adding in the ability to enter in champ stats.

At this time I don’t know if I want to separate this by season or do
this lifetime.

For now this will be a lifetime stats.

Goal is to look at a champ and evaluate
How often do I lose to a champ
How often do I win with a champ in the lineup
How often do I lose with a champ in the lineup